requirements.txt
langchain
langgraph
langchain-community
langchain-google-genai
chromadb
sentence-transformers
streamlit
python-dotenv
langchain-text-splitters
